Fix reload bug on delete / set captain
Due to the usage of <form>, when clicking you will submit the form.
But it has no action, so it won't do anything and just reloads.
See: https://stackoverflow.com/a/58763330

To fix this, we just remove the <form> tag and replace it with <div>.
The style of the page won't change and it won't cause this weird behavior.